move dynamic, object, ErrorResponse to kube-core
think this completes the type migration to kube-core
