+++ kube/src/api/mod.rs
@@ -20,18 +20,17 @@ mod subresource;
 pub use subresource::{AttachParams, Attachable, Executable};
 pub use subresource::{EvictParams, Evictable, LogParams, Loggable, ScaleSpec, ScaleStatus};
 
-pub(crate) mod object;
-pub use self::object::{Object, ObjectList, WatchEvent};
-
 // Re-exports from kube-core
-pub use kube_core::gvk::{GroupVersionKind, GroupVersionResource};
-pub use kube_core::dynamic::{ApiResource, DynamicObject};
-pub use kube_core::metadata::{ListMeta, ObjectMeta, Resource, ResourceExt, TypeMeta};
-pub use kube_core::request::Request;
 pub(crate) use kube_core::params;
+pub use kube_core::{
+    dynamic::{ApiResource, DynamicObject},
+    gvk::{GroupVersionKind, GroupVersionResource},
+    metadata::{ListMeta, ObjectMeta, Resource, ResourceExt, TypeMeta},
+    object::{Object, ObjectList, WatchEvent},
+    request::Request,
+};
 pub use params::{
     DeleteParams, ListParams, Patch, PatchParams, PostParams, Preconditions, PropagationPolicy,
 };
-
 
 #[cfg(feature = "admission")] pub mod admission;
